Fixes to placeholder positioning and animations
When non-current placeholders are moved to before the current
placeholder, it created problems because it moved the next item over.
This is solved by moving the item previous to the current placeholder
after the placeholder every time a non-current placeholder is moved bef
ore the current placeholder (except for when the target item is a
non-current placeholder).

Also added a new option `pointerVelocityThreshold` which prevents
triggering of _rearrange if the mouse velocity is below the threshold
(default is 0.5 pixels/ms). This prevents triggering lots of animations
when moving the mouse rapidly, which caused significant FPS drops.